AS : Moved the Example template test to the correct folder
