+++ frontend/tests/unit/Example.test.jsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { describe, it, expect } from "vitest";
-import Example from "./Example";
+import Example from "../../src/pages/Example/Example";
 
 
 //The Name of the page
